chore: rename task card component
